requirements.txt
torch
tensorboard
python-dotenv
dvc[all]
sacred
ruamel.YAML
git+https://github.com/Leonardo130709/ycutils.git